compiling

+++ aStar/node.h
@@ -8,9 +8,9 @@ private:
     Node *child; // next node towards goal
 
     // Costs of node
-    unsigned long long int g;
-    unsigned long long int h;
-    unsigned long long int f;
+    unsigned int g;
+    unsigned int h;
+    unsigned int f;
 public:
     Node(){
         parent = child = NULL;
@@ -23,8 +23,8 @@ public:
     void setParent(Node *p){ parent = p; }
     void setChild(Node *c){ child = c; }
 
-    void setG(unsigned long long int c){ g = c; }
-    void setH(unsigned long long int c){ h = c; }
+    void setG(unsigned int c){ g = c; }
+    void setH(unsigned int c){ h = c; }
     void setF(){ f =  g + h; }
 
     unsigned long long int getG(){ return g; }
